Fixing test errors

+++ termdim_windows.go
@@ -11,5 +11,5 @@ func getSize(fd int) (width, height int, err error) {
 	if err := windows.GetConsoleScreenBufferInfo(windows.Handle(fd), &info); err != nil {
 		return -1, -1, err
 	}
-	return uint16(info.Window.Right - info.Window.Left + 1), uint16(info.Window.Bottom - info.Window.Top + 1), nil
+	return info.Window.Right - info.Window.Left + 1, info.Window.Bottom - info.Window.Top + 1, nil
 }
